+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [IMPROVEMENT] By default, the XML encoding is not supported at all, and must be enabled explicitly. This also removes the dependency to the *lxml* library.
 - [IMPROVEMENT] 'find...()' methods are now part of the ResourceBase" class.
 - [FIX] Group resource creation fixed.
+- [IMPROVEMENT] Now using logging for debug output.
 
 ## Version 0.7 (2018-05-13)
 - [IMPROVEMENT] Added support for &lt;remoteCSE> resource type.